% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_respose_matrix.R
\name{AddNoise}
\alias{AddNoise}
\title{Add noise to response value}
\usage{
AddNoise(response.df, method)
}
\arguments{
\item{response.df}{A data frame. It contains the response data for one block
of drug combination screen.}

\item{method}{A characer. It refers to the method used for calculating noise.
Available options are "random" and "scale".}
}
\value{
A matrix. It contains the response value added with noises.
}
\description{
Function \code{AddNoise} calculates and add a noise to values in response
matrix. The methods used to calculate noises are: "random" and "scale".
}
\details{
There are two options for calculating noise in this function:
\itemize{
  \item \strong{random} The noises obey normal distribution ~N(0, 0.001).
  They are generated by fucntion \code{\link[stats]{rnorm}}.
  \item \strong{scale} The noises are generated by multiplying the total
  number of response values in matrix with 10^-10.
}
}
