% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_drug.R
\name{AnnotateDrug}
\alias{AnnotateDrug}
\title{Annotate drug}
\usage{
AnnotateDrug(drug_names)
}
\arguments{
\item{drug_names}{A vector of characters containing the names for the drugs}
}
\value{
A list with two data frames:
  \itemize{
    \item \strong{drug} It contain the basic information from PubChem,
      ChEMBL, and UniChem database.
    \item \strong{target} It contains the target information for durgs from
      DTC database.
  }
}
\description{
This function is a wrapper for multiple query functions. It will query
PubChem, ChEMBL, UniChem, and DTC to fetch information for drugs
}
\examples{
AnnotateDrug(c("aspirin", ""))
}
