% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_synergy_score.R
\name{CalculateHSA}
\alias{CalculateHSA}
\title{Calculate HSA synergy score}
\usage{
CalculateHSA(response.mat)
}
\arguments{
\item{response.mat}{A drug cobination dose-response matrix. It's column name
and row name are representing the concerntrations of drug added to column and
row, respectively. The values in matrix indicate the inhibition rate to cell
growth.}
}
\value{
A matrix with synergy score calculated via Highest Single Agent (HSA)
reference model.
}
\description{
\code{CalculateHSA} calculates the synergy score matrix for a block of
drug combination by using Highest Single Agent (HSA) reference model.
}
\details{
This model is a reference model for evaluating the interaction between two
drugs. The basic assumption of this model is "The reference effect of drug
combination is the maximal single drug effect". More details about this model
could be found in original publication:
\href{https://www.ncbi.nlm.nih.gov/pubmed/2692037}{(Berenbaum, 1989)}.
}
