% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sensitivity.R
\name{CalculateSens}
\alias{CalculateSens}
\title{Calculate sensitivity score}
\usage{
CalculateSens(df)
}
\arguments{
\item{df}{A data frame. It contains two variables:
\itemize{
  \item \strong{dose} the concentrations of drugs.
  \item \strong{response} the response of cell lines at crresponding doses.
  We use inhibition rate of cell line growth to measure the response.
}}
}
\value{
A number. The sensitivity score calculated from input dose-response
data
}
\description{
Function \code{CalculateSens} calculates cell line sensitivity to a drug or a
combination of drugs from dose response curve.
}
\details{
This function measures the sensitivity by calculating the Area Under Curve
(AUC) according to certurn dose response curve. The lower bouder is chosen at
the first non-zero concentration in the dose response data.
}
