% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_cellosaurus.R
\name{CellVersion}
\alias{CellVersion}
\title{Check the Cellosaurus data version.}
\usage{

  CellVersion(file = "ftp://ftp.expasy.org/databases/cellosaurus/cellosaurus.xml",
  ...)
}
\arguments{
\item{file}{A character. The xml file contains the Cellosaurus dataset.}

\item{...}{Other arguments.}
}
\description{
Cellosaurus publishes its dataset as an XML file which could be downloaded
from ftp://ftp.expasy.org/databases/cellosaurus/cellosaurus.xml. This
could be used to check the publish version of this file (by default) or a
XML downloaded before (by passing the "file" argument).
}
\examples{
# Check the online Cellosaurus database version.
CellVersion()

# Check the local Cellosaurus XML document version.
CellVersion(system.file("extdata", "cellosaurus.xml", package = "TidyComb"))

}
