% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_drugcomb.R
\name{CheckCell}
\alias{CheckCell}
\title{Check current status of DrugComb}
\usage{
CheckCell(test)
}
\description{
Get the updated csv files from DrugComb database:
cell_line: cellosaurus_accession, id
select * from drug into outfile '/var/lib/mysql-files/dr.csv' fields terminated by ',' enclosed by '"' lines terminated by '\n';
drug: id, cid
tissue: id, name (*)
disease: name, id (*)
study: id
}
