% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_drugcomb.R
\name{CheckDisease}
\alias{CheckDisease}
\title{Check diseases archive in DrugComb}
\usage{
CheckDisease(test)
}
\arguments{
\item{test}{A character vector. It contains IDs of diseases which would
be checked.}
}
\value{
A list contains 3 elements:
\itemize{
  \item \strong{n} It indicates that there are totally n diseases archived
  in DrugComb database.
  \item \strong{old} It lists out the ID of diseases that
  \emph{have been archived} in DrugComb.
  \item \strong{new} It lists out the ID of diseases that \emph{are not yet}
  in DrugComb.
}
}
\description{
Function \code{CheckDisease} compares disease IDs passed to \code{test}
argument with diseases archived in DrugComb (extracted to
\code{\link{drugcomb}} R data). It will return the total number of diseases
have been archived in DrugComb as well as which diseases are already or not
yet archived in DrugComb.
}
