% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_drugcomb.R
\name{CheckDrug}
\alias{CheckDrug}
\title{Check whether input drugs are archived in DrugComb}
\usage{
CheckDrug(cids)
}
\arguments{
\item{cids}{A character vector. It contains CIDs of drugs which would be
checked.}
}
\value{
A list contains 3 elements:
\itemize{
  \item \strong{n}: It indicates that there are totally n drugs archived in
  DrugComb database.
  \item \strong{old}: It lists the CIDs of drugs that
  \emph{have been archived} in DrugComb.
  \item \strong{new}: It lists the CIDs of drugs that \emph{are not yet}
  in DrugComb.
}
}
\description{
Function \code{CheckDrug} compares drug CIDs passed to \code{cids} argument
with CIDs archived in DrugComb (extracted to \code{\link{drugcomb}} R data).
It will return the total number of drugs have been archived in DrugComb as
well as which drugs are already or not yet archived in DrugComb.
}
\author{
ShuyuZheng \email{shuyu.zheng@helsinki.fi}
}
