% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{CheckTemplate}
\alias{CheckTemplate}
\title{Check the input template}
\usage{
CheckTemplate(template)
}
\arguments{
\item{template}{a data frame. It is the input data from users for analysis}
}
\value{
Non
}
\description{
\code{CheckTemplate} function check the template inputted from users to
ensure they are acceptable to downstream functions.
}
\examples{
data <- read.csv(system.file("template.csv", package = "TidyComb"),
                 stringsAsFactors = FALSE)
CheckTemplate(data)

# Make some messy to data

data_m <- data[, -2]
CheckTemplate(data_m)

}
