% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chembl.R
\name{ChemblVer}
\alias{ChemblVer}
\title{Check version of Chembl database}
\usage{
ChemblVer()
}
\value{
A  vector of named characters contains the update information of
Chembl.
}
\description{
Check version of Chembl database
}
\examples{
print(ChemblVer())
}
