% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_chembl.R
\name{ChemblVersion}
\alias{ChemblVersion}
\title{Check version of ChEMBL database}
\usage{
ChemblVersion()
}
\value{
A vector of named characters contains the update information of
ChEMBL.
}
\description{
By calling function \code{ChemblVersion} without any argument, you can check
the current version ChEMBL database.
}
\details{
\code{ChemblVersion} is a fuction wrapping the "Status" function provided by
\href{https://www.ebi.ac.uk/chembl/ws}{ChEMBL Web Services}.
}
\examples{
print(ChemblVersion())
}
