% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_synergy_score.R
\name{CorrectBaseLine}
\alias{CorrectBaseLine}
\title{Base line correction}
\usage{
CorrectBaseLine(response.mat, ...)
}
\arguments{
\item{response.mat}{A drug cobination dose-response matrix. It's column name
and row name are representing the concerntrations of drug added to column and
row, respectively. The values in matrix indicate the inhibition rate to cell
growth.}
}
\value{
A matrix which base line have been adjusted.
}
\description{
\code{CorrectBaseLine} adjusts the base line of drug combination
dose-response matrix up to positive values.
}
