% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_cellosaurus.R
\name{DownloadCellosaurus}
\alias{DownloadCellosaurus}
\title{Download the xml file for Cellosaurus database}
\usage{
DownloadCellosaurus(file)
}
\arguments{
\item{file}{A character value. It indicates the path and name of the file to
save the downloaded data.}
}
\description{
This function will download the xml file for Cellosaurus data from
https://ftp.expasy.org/databases/cellosaurus
}
