% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugcomb.R
\docType{data}
\name{drugcomb}
\alias{drugcomb}
\title{Identifiers in DrugComb}
\format{A list with 5 data frames:
\describe{
  \item{cell_line}{Contains 2 variables: cellosaurus_accession, id.}
  \item{drug}{Contains 2 variables: id, cid.}
  \item{disease}{Contains 2 variables: id, name.}
  \item{tissue}{Contains 2 variables: name, id.}
  \item{study}{Contains 1 variable: id.}
}}
\source{
url{https://drugcomb.fimm.fi/}
}
\usage{
drugcomb
}
\description{
A list containing the identifiers of drugs, diseases, tissues, cell_lines,
studies archived in DrugComb.
}
\details{
This is a list of identifiers of drugs, diseases, tissues, cell_lines,
studies archived in DrugComb. Used to check the records currently exist in
DrugComb and avoid generate duplicated entries.

The data is request from DrugComb database, using MySQL commands:

\code{SELECT cellosaurus_accession,id FROM cell_line
SELECT id,cid FROM drug
SELECT * FROM disease
SELECT * FROM tissue
SELECT id FROM study;}
}
\keyword{datasets}
