% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_score.R
\name{FittingSingleDrug}
\alias{FittingSingleDrug}
\title{Fitting single drug dose-response model}
\usage{
FittingSingleDrug(response, ...)
}
\arguments{
\item{response}{A matrix. It contains two columns:
\itemize{
  \item \strong{conc} The concentration of drugs added in experiment.
  \item \strong{response} The response of cell lines to drug with different
  concentrations.
}}

\item{...}{}
}
\value{
A list. It contains two elements:
\itemize{
  \item \strong{fitted} Fitted values.
  \item \strong{model} Fitted model.
}
}
\description{
Function \code{FittingSingleDrug} fits dose-response model by using
\code{\link[drc]{drm}} function.
}
\details{
Pre-fitting process:
1. Change the 0 value in concentration into 10^-10 to avoide raising error
when taking log.
2. If the variance of "response" values equal to 0, add 10^-10 to the last
"response" value.

Model choice:
First use "L.4" model to fit the raw data. If error or waring occurs, use
"LL.4" model to fit log(raw data).
}
