% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cell.R
\name{GenerateCell}
\alias{GenerateCell}
\title{Generate "cell_line" and "cell_id" tables}
\usage{
GenerateCell(acc, file)
}
\arguments{
\item{acc}{A vector of charactors contains cellosaurus accessions of
interested cell lines.}

\item{file}{The path to Cellosaurus XML file which was posted on
\url{https://web.expasy.org/cellosaurus/}}
}
\value{
A list with 2 data frames:
\itemize{
  \item \strong{cell_line} The cell_line table prepared for uploading.
  \item \strong{cell_id} The DrugComb IDs for new cell lines.
}
}
\description{
Generate "cell_line" and "cell_id" tables
}
\author{
Jing Tang \email{jing.tang@helsinki.fi}
Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
}
