% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_drug.R
\name{GenerateDrug}
\alias{GenerateDrug}
\title{Generate drug table and drug id index table}
\usage{
GenerateDrug(cids)
}
\arguments{
\item{cids}{A vector of characters contains interested drugs}
}
\value{
A list contains 2 data frame: one contains new drug's information,
 another one is drug id cid index table.
}
\description{
Generate drug table and drug id index table
}
\examples{
input <- utils::read.csv(system.file("extdata",
                              "template.csv",
                               package = "TidyComb"),
                  stringsAsFactors = FALSE)
cids <- na.omit(unique(c(input$drug_row_cid, input$drug_col_cid)))
drug <- GenerateDrug(cids)
}
