% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_cellosaurus.R
\name{GetAllCell}
\alias{GetAllCell}
\title{Loading a Cellosaurus XML dataset}
\usage{
GetAllCell()
}
\value{
An XMLNode containing all cell lines' information archieved in
Cellosaurus dataset.
}
\description{
Cellosaus published "cellosaurus.xml" contains 5 part of information:
"header", "cell-line-list", "publication-list", "copyright". (more
information in "ftp://ftp.expasy.org/databases/cellosaurus/cellosaurus.xsd")
This function will parse the XML file and extract "cell-line-list" node which
includes all information about cell lines as a XML document object.
}
\examples{
all.cell <- GetAllCell()
}
