% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus.R
\name{GetCell}
\alias{GetCell}
\title{Find matitching cell-lines}
\usage{
GetCell(node, type, id)
}
\arguments{
\item{node}{An XMLNodelist. It is the output of \code{\link{GetAllCell}}
function which contains all cell lines' information in Cellosaurus dataset.}

\item{type}{A charatcer. It indicate the type of \code{id}. It could be
"names", "accession".}

\item{id}{A vector of characters. It is the name or Cellosaurus accession of
cell lines that will be searched in Cellosaurus XML file.}
}
\value{
An XMLNodeSet containing matched cell lines in the dataset. If no
cell line is matched, a NULL list will be return.
}
\description{
This function searches the value of <name> tags in Cellosaurus XML document
to find entries that matches the names provided in the \code{names}
parameter.
}
\examples{
all.cells <- GetAllCell(system.file("extdata",
                                    "cellosaurus.xml",
                                     package = "TidyComb"))
cell.lines <- GetCell(all.cells, c("U87", "A549"))

}
