% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus.R
\name{GetCellInfo}
\alias{GetCellInfo}
\title{Extract cell line information}
\usage{
GetCellInfo(node)
}
\arguments{
\item{node}{An "XMLInternalElementNode" extracted from the Cellosaurus XML
file by either \code{\link{GetAllCell}} or \code{\link{GetCell}}}

\item{term}{A vector of character. The type of information you want to
extract. A valible values are: "synonyms", "accession", "disease",
"diseaseID", "name", "all".}
}
\value{
A data frame containing the information selected by \code{term}
extracted from the \code{node} item.
}
\description{
This function will extract the information required by DrugComb database (
including cell name, cell synonyms, disease, disease_id ) of the cell lines.
The input should be "XMLInternalElementNode" that contains the information
of the cell lines. It could be output of either  \code{\link{GetAllCell}} (
all Cellosaurus archived cell lines) or \code{\link{GetCell}} (cell lines
you're interested in).
}
\examples{
all.cells <- GetAllCell(system.file("extdata",
                                    "cellosaurus.xml",
                                     package = "TidyComb"))
cell.lines <- GetCell(all.cells, c("U87", "A549"))
cell.info <- GetCellInfo(cell.lines, "synonyms")

}
