% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_cellosaurus.R
\name{GetCellName}
\alias{GetCellName}
\title{Extract primary name and synonyms of one cell line.}
\usage{
GetCellName(node)
}
\arguments{
\item{node}{An \code{XMLInternalElementNode} with only one
\emph{cell-line-list} node extracted from \emph{Cellosaurus xml file}.}
}
\value{
A data frame contains two variables:
\itemize{
  \item \code{name} the primary name of cell line.
  \item \code{synonyms} synonyms of the cell line separated by semicolons.
}
}
\description{
\code{GetCellName} extract primary name and synonyms from only one
\emph{cell-line-list} node in Cellosaurus xml file.
}
\details{
This function extracts the primary name (as "name") and synonyms (as
"synonyms") of an \code{XMLIntervalElementNode} object containing
one \emph{cell-line-list} node extracted from Cellosaurus xml file.

If you'd like to extract information from multiple \emph{cell-line-list}
nodes, combining this function with \code{xpathSapply} or
\code{sapply} is recommanded.
}
