% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_cellosaurus.R
\name{GetCellNameInOne}
\alias{GetCellNameInOne}
\title{Extract primary name and synonyms of one cell line into one character.}
\usage{
GetCellNameInOne(node)
}
\arguments{
\item{node}{An \code{XMLInternalElementNode} with only one
\emph{cell-line-list} node extracted from \emph{Cellosaurus xml file}.}
}
\value{
A character contains all names (Primary and synonyms) of cell line.
}
\description{
\code{GetCellNameInOne} extract primary name and synonyms from only one
\emph{cell-line-list} node in Cellosaurus xml file and paste them into one
character.
}
\details{
This function extracts all names(primary name and synonyms) of an
\code{XMLIntervalElementNode} object containing one \emph{cell-line-list}
node extracted from Cellosaurus xml file. Then paste them into one character.

If you'd like to extract information from multiple \emph{cell-line-list}
nodes, combining this function with \code{xpathSapply} or \code{sapply} is
recommanded.
}
