% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_chembl.R
\name{GetChemblPhase}
\alias{GetChemblPhase}
\title{Get max clinical trial phase from ChEMBL ID}
\usage{
GetChemblPhase(ids, quiet = TRUE)
}
\arguments{
\item{ids}{A vector of characters contains the ChEMBL IDs of drugs.
\emph{Note:} the leading and tailing whitespaces are not allowed.}

\item{quiet}{A logical value. If it is \code{TRUE}, the error messages will
not show during runing the function.}
}
\value{
A data frame contains 2 columns:
 \item{chembl_id}{Inputted ChEMBL ID of drugs.}
 \item{chembl_phase}{Max clinical trial phase of the matched drugs.}
}
\description{
\code{GetChemblPhase} retrieves max clinical trial phase of drugs accordint
to ChEMBL ID.
}
\details{
\code{GetChemblPhase} queries ChEMBL database via
\href{https://www.ebi.ac.uk/chembl/ws}{ChEMBL Web Services}. The input drug
identifier is ChEMBL ID and the out put information is "max clinical trial
phase".
}
\examples{
drug.info <- GetChemblPhase("CHEMBL25")
}
\author{
Jing Tang \email{jing.tang@helsinki.fi}
Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
}
