% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_pubchem.R
\name{GetCid}
\alias{GetCid}
\title{Match CID according to other identifiers}
\usage{
GetCid(ids, type = NULL, quiet = TRUE)
}
\arguments{
\item{ids}{A vector of characters contains the identifiers of drugs you are
interested in.}

\item{type}{A character indicates the type of identifiers passed by \code{id}
argument. Available types are:
\itemize{
\item \strong{smiles} Identifiers of drugs in the simplified molecular-input
line-entry system (SMILES).
\item \strong{inchi} The IUPAC International Chemical Identifier (InChI).
\item \strong{inchikey} Standard InChIKey of the drugs.
\item \strong{formula} Molecule formula of the drugs.
\item \strong{name} Name for drugs. This type could be used for searching
synonyms, NCGC IDs, Chembl IDs, CAS or any other kind of identifiers.
\item \strong{sdf} The SDF
}}

\item{quiet}{A logical value. Describe wether or not show error message
during retrieving data.}
}
\value{
A data frame contains two columns:
\itemize{
  \item \code{id} The identifiers inputted by user.
  \item \code{cid} The matched CID.
}
}
\description{
\code{GetCid} matches CIDs of drugs according to user-provided identifiers.
}
\details{
This function using the PubChem API
(\href{https://pubchem.ncbi.nlm.nih.gov/pug_rest/PUG_REST.html}{PUG REST}
to seach mached drugs. Available
identifiers are: name, SID, SMILES, InChI, SDF, InChIKey, molecula formula.

\strong{API USAGE POLICY:}
Please note that PUG REST is not designed for very large volumes (millions)
of requests. We ask that any script or application not make more than 5
requests per second, in order to avoid overloading the PubChem servers. If
you have a large data set that you need to compute with, please contact us
for help on optimizing your task, as there are likely more efficient ways to
approach such bulk queries.
}
\examples{
GetCid("aspirin", "name", quiet = TRUE)
}
