% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unichem.R
\name{GetIds}
\alias{GetIds}
\title{Get other identifiers from InChIKey}
\usage{
GetIds(ids, type)
}
\arguments{
\item{ids}{A vector of characters. It contains identifiers of the drug.}

\item{type}{A character or an integer. It specifies the type of identifiers
passed to \code{ids}. Acceptable values are:
\enumerate{
  \item \strong{An integer} indicates the "src_id" that UniChem assigned to
   various sources. A list of valid "src_id" can be found either on the
   \href{https://www.ebi.ac.uk/unichem/ucquery/listSources}
   {UniChem sources page} or by using \code{\link{GetAllSrcIds}} function.
  \item \strong{inchikey} Standard InChIKey of the drugs.
}}
}
\value{
A data frame with first column is the identifier passed to \code{ids}
argument and following columns are indentifiers of all resources integrated
by UniChem.
}
\description{
\code{GetIds} matches the identifiers of
\href{https://www.ebi.ac.uk/unichem/ucquery/listSources}{UniChem sources},
including DrugBank, KEGG, Zinc...
}
\details{
This function retrieves the of drug
via \href{https://www.ebi.ac.uk/unichem/info/webservices}{UniChem Web
Services}.
}
\examples{
aspirin <- GetIds(ids = "BSYNRYMUTXBXSQ-UHFFFAOYSA-N", type = "inchikey")
}
