% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus.R
\name{GetNames}
\alias{GetNames}
\title{Extract primary name and synonyms of one cell line.}
\usage{
GetNames(node)
}
\arguments{
\item{node}{An XMLInternalElementNode with only one cell line's information
which was extracted from Cellosaurus xml file.}
}
\value{
A data frame contains two variables: name (the primary name of cell
line) and synonyms (all the synonyms of the cell line pasted into one
character and separated by ";")
}
\description{
This function extracts the primary name (as "name") and synonyms(as"synonyms"
) of one cell line from an XMLIntervalElementNode object extracted from
Cellosaurus xml file. Combining this function with \conde{\link[base]{apply}}
or \code{\link[base]{sapply}} can extract names information from an
XMLNodeSet object.
}
