% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_pubchem.R
\name{GetPubchemPro}
\alias{GetPubchemPro}
\title{Get properties of drugs}
\usage{
GetPubchemPro(cids)
}
\arguments{
\item{cids}{A vector of integer or character indicates the CIDs of drugs.}
}
\value{
A data frame contains 4 columns:
\itemize{
  \item \strong{CID} CID of drugs which is inputted to \code{cids} argument.
  \item \strong{InChIKey} Standard InChIKey of matched drugs.
  \item \strong{CanonicalSMILES} Standard Canonical SMILES of matched drugs.
  \item \strong{MolecularFormula} Molecular formula for matched drugs.
}
}
\description{
\code{GetPubchemPro} function retrieves the properties (InChIKey, Canonical
SMILES, and molecula formula) of drugs from PubChem database via
\href{https://pubchemdocs.ncbi.nlm.nih.gov/pug-rest$_Toc494865567}{PUG REST}.
}
\examples{
property <- GetPubchemPro(c(1,2,3,4))
}
