% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_pubmed.R
\name{GetPubmed}
\alias{GetPubmed}
\title{Get publication information from Pubmed}
\usage{
GetPubmed(pmids, tool = "TidyComb", email = NULL)
}
\arguments{
\item{pmids}{A vector of character/numeric containing PubMed ID of
publications}

\item{tool}{The name of application, default is "TidyComb"}

\item{email}{User's e-mail address}
}
\value{
A data frame containing 6 variables:
\itemize{
  \item \strong{pubmed_id} The PubMed ID of the publication passed to
  \code{pmid} argument.
  \item \strong{year} The year of the publication date.
  \item \strong{name} The first name of the first author.
  \item \strong{title} The title of the publication.
  \item \strong{journal} The full name of journal where the publication was
  reposted.
  \item \strong{doi} The Digital Object Identifier (DOI) of the publication.
}
}
\description{
\code{GetPubmed} extract information (pulication date, first name of first
author, title of the publication, full name of journal, doi) of publications
from Pubmed database.
}
\details{
\code{GetPubmed} extract publications information via API provided by PMC
database.

The argument \code{tool} and \code{email} are required for access
\href{https://www.ncbi.nlm.nih.gov/pmc/tools/developers/}{API}.
Following is the "Note about API usage":

The use of our APIs is entirely free, and doesn't require an API key, but
we ask that you please:
\itemize{
  \item{Do not make concurrent requests, even at off-peak times;}
  \item{Include two parameters that help to identify your service or
application to our servers:
    \enumerate{
      \item tool should be the name of the application, as a string value
      with no internal spaces.
      \item email should be the e-mail address of the maintainer of the tool,
      and should be a valid e-mail address.
    }
  }
}
}
\examples{
study <- GetPubmed(c(26551875, 15627163))
}
