% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_pubchem.R
\name{GetPubPhase}
\alias{GetPubPhase}
\title{Get max clinical trial phase from Pubchem}
\usage{
GetPubPhase(cids, quiet = TRUE)
}
\arguments{
\item{cids}{A vector of integer or character indicates the CIDs of drugs.}

\item{quiet}{A logical value. If it is \code{TRUE}, the error message will
shut down during retrieving process.}
}
\value{
A data frame contains 4 columns:
\itemize{
  \item \strong{CID} CID of drugs which is inputted to \code{cids} argument.
  \item \strong{phase} Max clinical trial phase of matched drugs.
}
}
\description{
\code{GetPubPhase} function retrieves the max clinical trial phase of drugs
from PubChem database via
\href{https://pubchemdocs.ncbi.nlm.nih.gov/pug-rest$_Toc494865567}{PUG REST},
accordint to CIDs.
}
\examples{
clinical.phase <- GetPubPhase(c(1,2,3,4))
}
