% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_pubchem.R
\name{GetPubSynonymFromName}
\alias{GetPubSynonymFromName}
\title{Get drug synonyms from PubChem by Drug names}
\usage{
GetPubSynonymFromName(names)
}
\arguments{
\item{names}{A vector of characters containing names of drugs for searching.}
}
\value{
A Data frame contains:
\itemize{
  \item \strong{input_name} The name inputted by user.
  \item \strong{synonyms} Synonyms retrieved from PubChem. Different synonyms
  are separated by ";".
}
}
\description{
\code{GetPubSynonymFromName} queries PubChem database with drug names via
(\href{https://pubchem.ncbi.nlm.nih.gov/pug_rest/PUG_REST.html}{PUG REST} and
retrieves synonyms of drugs.
}
\examples{
names <- c("Aspirin", "5-FU")
synonyms <- GetPubNames(names)

}
