% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_DTC.R
\name{GetTarget}
\alias{GetTarget}
\title{Match target according to CHEMBL IDs}
\usage{
GetTarget(chembl_ids)
}
\arguments{
\item{chembl_ids}{A vector of characters contains the CHEMBL IDs of drugs}
}
\value{
A data frame contains two columns:
\itemize{
  \item \code{chembl_id} Input CHEMBL ID
  \item \code{target_gene_name} Gene name for the target
  \item \code{inhibitor_type} The type of the inhibitor
  \item \code{target_protein_class} The protein class for target
  \item \code{target_organism} The organism used to test the target activity
  \item \code{target_protein_name} The protein name for target
}
}
\description{
This function matches target of drugs according to user-provided identifiers.
}
\details{
This function queries DTC database via
\href{chrome-extension://oemmndcbldboiebfnladdacbdfmadadm/file:///Users/shuzhe/Desktop/Api_documentation.pdf}{API}
to query targets with CHEMBL ID
}
\examples{
GetTarget("CHEMBL25")
}
\author{
Jing Tang \email{jing.tang@helsinki.fi}
Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
}
