% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unichem.R
\name{GetUnichem}
\alias{GetUnichem}
\title{Extract information from Chembl database}
\usage{
GetUnichem(ids, type)
}
\arguments{
\item{ids}{A vector of characters. It contains identifiers of the drug.}

\item{type}{A character. It specifies the type of identifiers passed to
\code{ids}. Acceptable inputs are:
\enumerate{
  \item src_compound_id: It is the individual compound identifier provided by
each of the sources.
  \item src_id: It is integers that represent the various sources in UniChem.
A list of valid src_id's can be found either on the
\href{https://www.ebi.ac.uk/unichem/ucquery/listSources}
{UniChem sources page} or by using \code{\link{GetAllSrcIds}} function.
  \item InChIKey:
}}
}
\value{

}
\description{
This function retrieves the Chembl ID and max clinical trial phase of drug
via \href{https://www.ebi.ac.uk/unichem/info/webservices}{UniChem Web
Services}.
}
