% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_own_rank.R
\name{own_rank}
\alias{own_rank}
\title{Rank the vectors}
\usage{
own_rank(x)
}
\arguments{
\item{x}{a vector need to be ranked}
}
\value{
a vector contains the rank of original vector
}
\description{
Rank the vectors
}
