% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pipeline.R
\name{ParCalculateTemplate}
\alias{ParCalculateTemplate}
\title{Parallel Calculate Drug Combination data in template format}
\usage{
ParCalculateTemplate(template, ...)
}
\arguments{
\item{template}{a dataframe in the format as template. Columns "block_id",
"drug_row", "drug_col", "response", "conc_r", "conc_c", "conc_r_unit",
"conc_c_unit","cell_line_name", "drug_row", "drug_col" are reqired.}

\item{...}{Other arguments required by nested functions}
}
\value{
A list. It contains 4 tables:
  \itemize{
    \item \strong{synergy} It contains the modified response value and 4
    type of synergy scores of each drug dose response pair.
    \item \strong{summary} It contains summarized information of each
    blocks: synergy scores, css, dss, S
    \item \strong{curve} It contains the coefficients from single drug dose
    response curve.
    \item \strong{surface} It contains the smoothed response value and
    synergy scores of each drug dose response pair.
 }
}
\description{
Parallel Calculate Drug Combination data in template format
}
