% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_pipeline.R
\name{pipeline}
\alias{pipeline}
\title{The pipeline for calculation}
\usage{
pipeline(response)
}
\arguments{
\item{response}{A data frame. It must contain variables: "block_id",
"drug_row", "drug_col", "response", "conc_r", "conc_c", "conc_r_unit",
 "conc_c_unit","cell_line_name"}
}
\value{
A list contains 5 data frames: "response_with_scores", "response",
"summary", "surface", "curve".
}
\description{
\code{pipeline} mainly using the \code{response} value in \code{template}
file to calculate drug combination effect related scores, including IC50,
synergy scores, CSS...
}
\details{
\code{pipeline} function takes a data frame as input which must contain
variables: "block_id", "drug_row", "drug_col", "response", "conc_r", "conc_c",
"conc_r_unit", "conc_c_unit","cell_line_name".

It will calculate and generate following 5 tables:
\enumerate{
  \item \strong{response_with_scores}: Synergy scores calculated in
  4 methods (ZIP, Bliss, Loewe, and HSA) via calling function
  \code{\link{GenerateScore}}.
  \item \strong{response_out}: "response" table in the form required by
  "Drug Comb" database.
  \item \strong{summary}: IC50, DSS, CSS scores for each blocks calculated by
  calling function \code{\link{GenerateSummary2}}.
  \item \strong{surface}: "surface" table for visualization generated by
  calling function \code{\link{GenerateSurface3}}.
  \item \strong{curve}: "curve" table for uploading to Drug Comb generated by
  calling function \code{\link{GenerateCurve}}
}
}
