% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_smoothing.R
\name{smoothing}
\alias{smoothing}
\title{Smoothing the dose response or synergy score surface}
\usage{
smoothing(scores.mat, len = 2)
}
\arguments{
\item{scores.mat}{matrix, the matix for response valuses or synergy scores.}

\item{len}{integer, the number of data points will be added between two
neighbor dosages in original matrix.}
}
\value{
The matrix with addational data points.
}
\description{
This function will add the data points into the existing score matrix to make
the surface in 3D visualization smoothly. Kriging algotithom is used for
imputation.
}
\author{
Jing Tang \email{jing.tang@helsinki.fi}
Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
}
