% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyComb.R
\docType{package}
\name{TidyComb}
\alias{TidyComb}
\alias{TidyComb-package}
\title{TidyComb: Preparing drug combination data for DrugComb}
\description{
This package wrapped the functions that could be used to tidy drug
combination data from various sources into the format that is required by
uploading to DrugComb database.
}
\details{
The TidyComb package provides three categories of important
functions :
\enumerate{
  \item Retrieving data from PubChem and Chembl, and Cellosaurus databases;
  \item calculating drug synergy scores and curves;
  \item Calculating drug combination effect surface.
}
}
\section{Retrieving data from databases}{

The following functions are provided to
}

