% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_cellosaurus.R
\name{UpdateCell}
\alias{UpdateCell}
\title{Update local Cellosaurus XML file}
\usage{
UpdateCell(file)
}
\arguments{
\item{file}{The path to the local Cellosaurus file.}
}
\value{
A message about the version checking results. If the local file is
\strong{not} up-to-date, the local file will be updated with online
Cellosaurus data.
}
\description{
\code{UpdateCell} checks and compares the versions of online and
local Cellosaurus XML file. If the local file is not up-to-date, it will
automatically update the local file.
}
\details{
It is \strong{not} recommended to directly parse the online Cellosaurus
dataset. Because it will cost a large amount of memory which might leads to
a crash of R. It is better to download the dataset and then parse the local
Cellosaurus xml file. In this case, it is necessary to update your local file
before parsing it.

\code{UpdateCell} using the \code{\link{CellVersion}} function to check the
current online version of Cellosaurus dataset and the version of a local
file. It will automatically update the local file if it is not up-to-date.

Warning: The local file will be overwritted when updating data. If you'd like
to keep the old version of dataset, please backup it before updating.
}
\examples{
UpdateCell(system.file("extdata", "cellosaurus.xml", package = "TidyComb"))
}
