% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_pubchem.R
\name{UpdateCid}
\alias{UpdateCid}
\title{Get "Preferred Compound" of "Non-live" compounds in PubChem}
\usage{
UpdateCid(cids)
}
\arguments{
\item{cids}{a vector contains the CIDs of drugs on which test will be done.}
}
\value{
a data frame contains 2 columns:
  \itemize{
    \item \strong{old_cid} The inputted CIDs.
    \item \strong{new_cid} If the tested compound is "Non-live", it is the
    CID of "preferred compound". If the tested compound is still "live", it
    is \code{NA}.
  }
}
\description{
Fuction \code{UpdateCid} takes CIDs as input. If inputted compounds was
labeled as "Non-live" in PubChem, the CID of the "preferred Compound" will
be returned. If the inputted CID is still "live" in PubChem, \code{NA} will
be returned.
}
\examples{
res <- UpdateCid(1:10)
}
\author{
Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
}
