% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_score.R
\name{CalculateBliss}
\alias{CalculateBliss}
\title{Calculate Bliss synergy score}
\usage{
CalculateBliss(response.mat, Emin = NA, Emax = NA)
}
\arguments{
\item{response.mat}{A drug cobination dose-response matrix. It's column name
and row name are representing the concerntrations of drug added to column and
row, respectively. The values in matrix indicate the inhibition rate to cell
growth.}

\item{Emin}{A numeric or \code{NA}. It specifies the minimum value in the
fitted dose-response curve. Default setting is \code{NA}.}

\item{Emax}{A numeric or \code{NA}. It specifies the maximum value in the
fitted dose-response curve. Default setting is \code{NA}.}

\item{correction}{A logical value. It indicates whether \emph{baseline
correction} needed before calculation. Default value is \code{TRUE}, which
means the correction will be done.}
}
\value{
A matrix with  \eqn{\Delta} score calculated via Zero Interaction
Potency (ZIP) method
}
\description{
\code{CalculateBliss} calculates the \eqn{\Delta} score matrix for a block of
drug combination by using a reference druginteraction model introduced by .
}
\details{
zero interaction potency (ZIP) is a reference model for evaluating the
interaction between two drugs. It captures the drug interaction relationships
by comparing the change in the potency of the dose-response curves between
individual drugs and their combinations. More details about this model could
be found in original publication
\href{https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1744-7348.1939.tb06990.x}{(Bliss, 1939)}.
}
