% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_synergy_score.R
\name{CalculateLoewe}
\alias{CalculateLoewe}
\title{Calculate Loewe synergy score}
\usage{
CalculateLoewe(response.mat, quiet = TRUE, ...)
}
\arguments{
\item{response.mat}{A drug cobination dose-response matrix. It's column name
and row name are representing the concerntrations of drug added to column and
row, respectively. The values in matrix indicate the inhibition rate to cell
growth.}

\item{quiet}{A logical value. If it is \code{TRUE} then the warning message
will not show during calculation.}
}
\value{
A matrix with synergy score calculated via reference model introduced
by C. I. Bliss.
}
\description{
\code{CalculateLoewe} calculates the synergy score matrix for a block of
drug combination by using a druginteraction reference model introduced by
Loewe in 1953.
}
\details{
This model is a reference model for evaluating the interaction between two
drugs. The basic assumption of this model is "The referece effect of drug
combination is the expected effect of a drug combined with itself". More
details about this model could be found in original publication:
\href{https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1744-7348.1939.tb06990.x}{(Bliss, 1939)}.
}
