% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pipeline.R
\name{CalculateMat}
\alias{CalculateMat}
\title{Chain all calculation about one dose-response matrix}
\usage{
CalculateMat(response.mat, noise = TRUE, correction = "non", ...)
}
\arguments{
\item{response.mat}{A matrix contain the drug combination reaponse value.
Its column names are doses of drug added along columns. Its row name are
doses of drug added along rows. \cr
\strong{Note}: the matrix should be sorted by: 1. The concentrations along
the column increase \emph{from left to right}; 2. The concentrations along
the row increase \emph{from top to bottom}.}

\item{noise}{a logical value. It indicates whether or not adding noise to
to the "response" values in the matrix. Default is \code{TRUE}.}

\item{correction}{a string. It indicates which method used by function
\code{\link{CorrectBaseLine}} for base line correction.
  \itemize{
    \item \code{non}: no baseline correction;
    \item \code{par}: only correct base line on negative values in the matrix;
    \item \code{all}: correct base line on all the values in the matrix.
  }}

\item{...}{Other argumants required by nested functions. Some important
arguments are:
\itemize{
   \item \code{Emin} and \code{Emax} inherited from function
         \code{link{FitDoseResponse}}.
}}
}
\value{
It contains 4 tables:
  \itemize{
    \item \strong{synergy} It contains the modified response value and 4
    type of synergy scores of each drug dose response pair.
    \item \strong{summary} It contains summarized information of each
    blocks: synergy scores, css, ri, S
    \item \strong{curve} It contains the coefficients from single drug dose
    response curve.
    \item \strong{surface} It contains the smoothed response values and
    synergy scores for plotting the scores' landscape.
 }
}
\description{
\code{CalculateMat} chains all calculations about one dose-response matrix (
one drug-drug interaction block) together. The calculations includes:
dose-response curve fitting, synergy scores (ZIP, Bliss, Loewe,
HSA), drug sensitivity (RI, CSS), generate drug-drug response surface and
generating summary scores for each block.
}
\details{
The steps for calculation:
\enumerate{
  \item Pre-process Matrix
    \enumerate{
      \item Impute for missing values (with the average of values from the
          nearest four cells) in original matrix by using function
          \code{\link{ImputeNear}}.
      \item Add noise(A small random number ranging from 0 to 0.001) to
          original matrix by using function \code{line{AddNoise}}.
      )
      \item Correct baseline using function \code{correction} with the method
      selected by parameter \code{correction}.
    }
  \item Single drug process
    \enumerate{
      \item Extract and fitting single drugs.
      \item Extract coeficients from fitted model. (b, c, d, e, IC50)
      \item Calculate RI
    }
  \item Whole response matrix process
    \enumerate{
      \item Calculate Synergy Scores
      \item Calculate Surface
      \item Calculate CSS
    }
  \item Summarize and generate surface
}
}
\examples{
data <- read.csv(system.file("template.csv", package = "TidyComb"),
                 stringsAsFactors = FALSE)
response.mat <- reshape2::acast(conc_r~conc_c, value.var = "response",
                                data = data[data$block_id == 1, ])
res <- CalculateMat(response.mat)
}
\author{
Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
}
