% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pipeline.R
\name{CalculateMat}
\alias{CalculateMat}
\title{Chain all calculatrion about one dose-response matrix together}
\usage{
CalculateMat(response.mat, impute = TRUE, noise = TRUE, ...)
}
\arguments{
\item{response.mat}{A matrix contain the drug combination reaponse value.
Its column names are doses of drug added along columns. Its row name are
doses of drug added along rows.}

\item{impute}{A logical function. It indicates whether or not imputing data
to filling the "NA" values. Defult is \code{TRUE}.}

\item{noise}{a logical value. It indicates whether or not adding noise to
to the "response" values in the matrix. Default is \code{TRUE}.}

\item{...}{Other argumants required by nested functions. Some important
arguments are:
\itemize{
   \item \code{method} inherited from function \code{\link{CorrectBaseLine}};
   \item \code{Emin} and \code{Emax} inherited from function
         \code{FitDoseResponse}.
}}
}
\value{
It contains 4 tables:
  \itemize{
    \item \strong{synergy} It contains the modified response value and 4
    type of synergy scores of each drug dose response pair.
    \item \strong{summary} It contains summarized information of each
    blocks: synergy scores, css, dss, S
    \item \strong{curve} It contains the coefficients from single drug dose
    response curve.
    \item \strong{surface} It contains the smoothed response value and
    synergy scores of each drug dose response pair.
 }
}
\description{
\code{CalculateMat} chains all calculations about one dose-response matrix (
one drug-drug interaction block) together. The calculations includes:
dose-response curve fitting, synergy scores (ZIP, Bliss, Loewe,
HSA), drug sensitivity (DSS, CSS), generate drug-drug response surface and
generating summary scores for each block.
}
\details{
The steps for calculation:
\enumerate{
  \item Pre-process Matrix
    \enumerate{
      \item Impute for missing value in original matrix.
      \item Add noise to original matrix.
      \item Correct baseline to 0, if \code{correction} is \code{TRUE}.
    }
  \item Single drug process
    \enumerate{
      \item Extract and fitting single drugs.
      \item Extract coeficients from fitted model. (b, c, d, e, IC50)
      \item Calculate DSS
    }
  \item Whole response matrix process
    \enumerate{
      \item Calculate Synergy Scores
      \item Calculate Surface
      \item Calculate CSS
    }
  \item Summarize and generate surface
}
}
