% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_chain.R
\name{CalculateMat}
\alias{CalculateMat}
\title{Chain all calculatrion about one dose-response matrix together}
\usage{
CalculateMat(response.mat, correction = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
\code{CalculateMat} chains all calculations about one dose-response matrix (
one drug-drug interaction block) together. The calculations includes:
dose-response curve fitting, synergy scores (ZIP, Bliss, Loewe,
HSA), drug sensitivity (DSS, CSS), generate drug-drug response surface and
generating summary scores for each block.
}
\details{
The steps for calculation:
\enumerate{
  \item Pre-process Matrix
    \enumerate{
      \item Impute for missing value in original matrix.
      \item Add noise to original matrix.
      \item Correct baseline to 0, if \code{correction} is \code{TRUE}.
    }
  \item Single drug process
    \enumerate{
      \item Extract and fitting single drugs.
      \item Extract coeficients from fitted model. (b, c, d, e, IC50)
      \item Calculate DSS
    }
  \item Whole response matrix process
    \enumerate{
      \item Fit dose-response model through whole matrix
      \item Calculate CSS
      \item Calculate Synergy Scores
    }
  \item Summarize and generate surface
}
}
