% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pipeline.R
\name{CalculateTemplate}
\alias{CalculateTemplate}
\title{Calculate Drug Combination data in template format}
\usage{
CalculateTemplate(template, debug = FALSE, ...)
}
\arguments{
\item{template}{a dataframe in the format as template. Columns "block_id",
"drug_row", "drug_col", "response", "conc_r", "conc_c", "conc_r_unit",
"conc_c_unit","cell_line_name", "drug_row", "drug_col" are reqired.}

\item{debug}{a logical value. If it is \code{TRUE}, block ID will be printed
to console. The default setting is \code{FALSE}}

\item{...}{Other arguments required by nested functions. Some important
arguments are:
 \itemize{
   \item \code{impute} and \code{noise} inherited from function
         \code{CalculateMat};
   \item \code{method} inherited from function \code{CorrectBaseLine};
   \item \code{Emin} and \code{Emax} inherited from function
         \code{FitDoseResponse}.
}}
}
\value{
A list. It contains 4 tables:
  \itemize{
    \item \strong{synergy} It contains the modified response value and 4
    type of synergy scores of each drug dose response pair.
    \item \strong{summary} It contains summarized information of each
    blocks: synergy scores, css, dss, S
    \item \strong{curve} It contains the coefficients from single drug dose
    response curve.
    \item \strong{surface} It contains the smoothed response value and
    synergy scores of each drug dose response pair.
 }
}
\description{
Calculate Drug Combination data in template format
}
\author{
Shuyu Zheng{shuyu.zheng@helsinki.fi}
}
