% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_drugcomb.R
\name{CheckCell}
\alias{CheckCell}
\title{Check cells archive in DrugComb}
\usage{
CheckCell(test)
}
\arguments{
\item{test}{A character vector. It contains \emph{cellosaurus accessions}
of cell lines which would be checked.}
}
\value{
A list contains 3 elements:
\itemize{
  \item \strong{n} It indicates that there are totally n cell lines archived
  in DrugComb database.
  \item \strong{old} It lists out the \emph{cellosaurus accessions} of cell
  lines that \emph{have been archived} in DrugComb.
  \item \strong{new} It lists out the \emph{cellosaurus accessions} of cell
  lines that \emph{are not yet} in DrugComb.
}
}
\description{
Function \code{CheckCell} compares cellosaurus accessions passed to
\code{test} argument with cell lines archived in DrugComb (extracted to
\code{\link{drugcomb}} R data). It will return the total number of cell lines
have been archived in DrugComb as well as which cell lines are already or not
yet archived in DrugComb.
}
