% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_drugcomb.R
\name{CheckDisease}
\alias{CheckDisease}
\title{Check whether input diseases are archived in DrugComb}
\usage{
CheckDisease(disease_df)
}
\arguments{
\item{disease_df}{A data frame . It contains tow columns:
\itemize{
  \item \strong{id}: the NCIt ID of the disease.
  \item \strong{name}: the name of the disease.
}}
}
\value{
A list contains 3 elements:
\itemize{
  \item \strong{n}: It indicates that there are totally n diseases archived
  in DrugComb database.
  \item \strong{old}: It lists the names and IDs of diseases that
  \emph{have been archived} in DrugComb.
  \item \strong{new}: It lists the names and IDs of diseases that
  \emph{are not yet} in DrugComb.
}
}
\description{
Function \code{CheckDisease} compares disease IDs passed to \code{test}
argument with diseases archived in DrugComb (extracted to
\code{\link{drugcomb}} R data). It will return the total number of diseases
have been archived in DrugComb as well as which diseases are already or not
yet archived in DrugComb.
}
\author{
ShuyuZheng \email{shuyu.zheng@helsinki.fi}
}
