% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_template.R
\name{CheckTemplate}
\alias{CheckTemplate}
\title{Check the variables of "template" file}
\usage{
CheckTemplate(template)
}
\arguments{
\item{template}{The template table which is provided by user for uploading
data to DrugComb database.}
}
\value{
If the template lack some of required elements, the program will be
shut down with an error. If the the template is all right, the program will
continue with a message tell user the template is OK.
}
\description{
Function \code{CheckTemplate} checks the variables in input \code{template}
to see whether all required elements have been provided.
}
\details{
The \code{template} file must have following variables:
\itemize{
  \item \strong{block_id} Identifier for each drug combination reaction
  block.
  \item \strong{drug_row} Identifier for drugs added to rows in certain
  block.
  \item \strong{drug_col} Identifier for drugs added to columns in certain
  block.
  \item \strong{response} Response value (inhibition rate for cell growth or
  other
  type of signals) detected from certain well in test blocks.
  \item \strong{conc_r} Concentration of drugs added to rows in certain
  block.
  \item \strong{conc_c} Concentration of drugs added to columns in certain
  block.
  \item \strong{conc_r_unit} The unit used to measure concentration of drugs
  added in rows.
  \item \strong{conc_c_unit} The unit used to measure concentration of drugs
  added in columns.
  \item \strong{cell_line_name} The name of cell line in which drug
  combination screen was performed.
}
}
