% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_chembl.R
\name{ChemblVersion}
\alias{ChemblVersion}
\title{Check version of Chembl database}
\usage{
ChemblVersion()
}
\value{
A vector of named characters contains the update information of
Chembl.
}
\description{
Check version of Chembl database
}
\examples{
print(ChemblVersion())
}
