% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dose_response_model.R
\name{FitDoseResponse}
\alias{FitDoseResponse}
\title{Fitting single drug dose-response model}
\usage{
FitDoseResponse(data, Emin = NA, Emax = NA)
}
\arguments{
\item{data}{A data frame. It contains two columns:
\itemize{
  \item \strong{conc} The concentration of drugs added in experiment.
  \item \strong{response} The response of cell lines to drug with different
  concentrations.
}}

\item{Emin}{A numeric or \code{NA}. It specifies the minimum value in the
fitted dose-response curve. Default setting is \code{NA}.}

\item{Emax}{A numeric or \code{NA}. It specifies the maximum value in the
fitted dose-response curve. Default setting is \code{NA}.}
}
\value{
An object of class 'drc'. It contains imformation of fitted model.
}
\description{
Function \code{FitDoseResponse} fits dose-response model by using
\code{\link[drc]{drm}} function.
}
\details{
Pre-fitting process:
1. Change the 0 value in concentration into 10^-10 to avoide raising error
when taking log.
2. If the variance of "response" values equal to 0, add 10^-10 to the last
"response" value.

Model choice:
First use "L.4" model to fit the raw data. If error or waring occurs, use
"LL.4" model to fit log(raw data).
}
