% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cell_tissu_disease.R
\name{GenerateCell}
\alias{GenerateCell}
\title{Generate "cell_line" and "tissue" file}
\usage{
GenerateCell(acc)
}
\arguments{
\item{acc}{A vector of charactors contains cellosaurus accessions of
interested cell lines.}
}
\value{
A list with 3 data frame contains
\itemize{
  \item \strong{cell_line} The cell_line table prepared for uploading.
  \item \strong{tissue} The tissue table prepared for uploading.
  \item \strong{disease} The disease table prepared for uploading.
}
}
\description{
Generate "cell_line" and "tissue" file
}
