% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_cellosaurus.R
\name{GetAccession}
\alias{GetAccession}
\title{Extract the Cellosaurus accession ID of one cell line.}
\usage{
GetAccession(node)
}
\arguments{
\item{node}{An XMLInternalElementNode with only one cell line's information
which was extracted from Cellosaurus xml file.}
}
\value{
A character. The Cellosaurus accession ID of cell line
}
\description{
This function extract Cellosaurus accession ID of one cell line from an
XMLIntervalElementNode object extracted from Cellosaurus xml file. Combining
this function with \code{\link[base]{apply}} or \code{\link[base]{sapply}}
can extract disease information from an XMLNodeSet object.
}
\examples{
node <- GetAllCell(system.file("extdata", "cellosaurus.xml", package = "TidyComb"))
cell <- GetCell(node, c("U251", "U87"), "name")

# get Cellosaurus Accession for first cell line
accession <- GetAccession(cell[[1]])

# get Cellosaurus Accession for all cell lines
accessions <- sapply(cell, GetAccession)
}
