% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_unichem.R
\name{GetAllSrcIds}
\alias{GetAllSrcIds}
\title{Get all src_ids}
\usage{
GetAllSrcIds()
}
\value{
A data frame contains 3 columns:
\item{name}{Name of the resources.}
\item{src_id}{"src_id" of the resources.}
\item{description}{A brief description of the content of the source.}
}
\description{
Obtain all resources and corresponding \code{src_ids} currently in UniChem
}
\examples{
resource <- GetAllSrcIds()
}
\author{
Jing Tang \email{jing.tang@helsinki.fi}
Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
}
