% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_cellosaurus.R
\name{GetCell}
\alias{GetCell}
\title{Find matching cell-lines}
\usage{
GetCell(node, ids, type = "name")
}
\arguments{
\item{node}{An XMLNodelist. It is the output of \code{\link{GetAllCell}}
function which contains all cell lines' information in Cellosaurus dataset.}

\item{ids}{A vector of characters. It is the name or Cellosaurus accession of
cell lines that will be searched in Cellosaurus XML file.}

\item{type}{A charatcer. It indicate the type of \code{id}. It could be
"name", "accession".}
}
\value{
An XMLNodeSet containing matched cell lines in the dataset. If no
cell line is matched, a NULL list will be return.
}
\description{
This function searches the value of <name> tags in Cellosaurus XML document
to find entries that matches the names provided in the \code{names}
parameter.
}
\examples{
node <- GetAllCell(file = system.file("extdata", "cellosaurus.xml",
                                      package = "TidyComb"))
cell.lines <- GetCell(node, c("U87", "A549"), "name")

}
