% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus.R
\name{GetCellInfo}
\alias{GetCellInfo}
\title{Extract cell line information}
\usage{
GetCellInfo(node)
}
\arguments{
\item{node}{An "XMLInternalElementNode" extracted from the Cellosaurus XML
file by either \code{\link{GetAllCell}} or \code{\link{GetCell}}}
}
\value{
A data frame contains 5 variables are:
  \item{name}{primary name of cell line.}
  \item{synonyms}{synonyms of cell lines separated by semicolons.}
  \item{accession}{Cellosaurus Accession ID for cell lines.}
  \item{disease}{diseases that are associated with the cell lines.}
  \item{disease_id}{NCI Thesaurus entry code of the disease.}
}
\description{
This function will extract the primary name, synonyms, Cellosausurs Accession
ID, disease, disease_id) of the cell lines and Wrap them into one data frame.
If you prefer some not all of these data, \code{\link{GetNames}},
\code{\link{GetDisease}}, or \code{\link{GetAccession}} are recommended.
}
\examples{
all.cells <- GetAllCell(system.file("extdata",
                                    "cellosaurus.xml",
                                     package = "TidyComb"))
cell.lines <- GetCell(all.cells, c("U87", "A549"))
cell.info <- GetCellInfo(cell.lines)

}
