% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_cellosaurus.R
\name{GetCellInfo}
\alias{GetCellInfo}
\title{Extract cell line information}
\usage{
GetCellInfo(accessions, parsed_cell_file)
}
\arguments{
\item{accessions}{A vector containing the Cellosaurus accessions for
searching}

\item{parsed_cell_file}{An "XMLInternalElementNode" extracted from the
Cellosaurus XML file by function \code{\link{ParseCell}}}
}
\value{
A data frame containing 6 columns:
\itemize{
  \item \strong{name} primary name of cell lines.
  \item \strong{synonyms} the synonyms of cell lines.
  \item \strong{cellosaurus_accession} Cellosaurus Accession for cell lines.
  \item \strong{disease_name} diseases that are associated with the cell
    lines
  \item \strong{disease_id} NCI Thesaurus entry code of the diseases.
  \item \strong{tissue} Tissues that the cell line generated from. Following
  the CCLE category.
}

cell line information selected by \code{info}
}
\description{
This function will extract the primary name, synonyms, Cellosausurs Accession
ID, disease, disease_id, tissue) of the inputted cell lines and wrap them
into a data frame.
}
\author{
Jing Tang \email{jing.tang@helsinki.fi}
Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
}
