% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_cellosaurus.R
\name{GetCellInfo}
\alias{GetCellInfo}
\title{Extract cell line information}
\usage{
GetCellInfo(node, info = "accession")
}
\arguments{
\item{node}{An "XMLInternalElementNode" extracted from the Cellosaurus XML
file by either \code{\link{GetAllCell}} or \code{\link{GetCell}}}

\item{info}{A Character indicate about what kind of infomation you want to
extract. Available values are:
\itemize{
  \item \strong{name} primary name and synonyms of cell lines.
  \item \strong{accession} Cellosaurus Accession ID for cell lines.
  \item \strong{disease} diseases that are associated with the cell lines and
   corresponding NCI Thesaurus entry code.
  \item \strong{tissue} Tissues that the cell line generated from. Following the CCLE
   category.
}}
}
\value{
A data frame contains cell line information selected by \code{info}
}
\description{
This function will extract the primary name, synonyms, Cellosausurs Accession
ID, disease, disease_id) of the cell lines and Wrap them into one data frame.
If you prefer some not all of these data, \code{\link{GetNames}},
\code{\link{GetDisease}}, or \code{\link{GetAccession}} are recommended.
}
\examples{
all.cells <- GetAllCell(system.file("extdata",
                                    "cellosaurus.xml",
                                     package = "TidyComb"))
cell.lines <- GetCell(all.cells, c("U87", "A549"), "name")
cell.info <- GetCellInfo(cell.lines)

}
