% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_pubchem.R
\name{GetCid}
\alias{GetCid}
\title{Match CID according to other identifiers}
\usage{
GetCid(ids, type, quiet = TRUE)
}
\arguments{
\item{ids}{A vector of characters contains the identifiers of drugs for
searching.}

\item{type}{A character indicates the type of identifiers passed to \code{id}
argument. Available types are:
\itemize{
\item \strong{smiles} Identifiers of drugs in the simplified molecular-input
line-entry system (SMILES).
\item \strong{inchikey} Standard InChIKey of the drugs.
\item \strong{name} Name for drugs. This type could be used for searching
synonyms, NCGC IDs, Chembl IDs, CAS or any other kind of identifiers.
}}

\item{quiet}{A logical value. If it is \code{TRUE}, the error message
during retrieving data will not show in console.}
}
\value{
A data frame contains two columns:
\itemize{
  \item \code{input_id} The identifiers input by user.
  \item \code{cid} The matched CID.
}
}
\description{
\code{GetCid} matches CID of drugs according to user-provided identifiers.
}
\details{
\code{GetCid} queries PubChem database via
\href{https://pubchem.ncbi.nlm.nih.gov/pug_rest/PUG_REST.html}{PUG REST}
to search matched CIDs of drugs according to other identifiers. Available
identifiers are: name, SMILES, InChIKey.

Following is the \strong{"API USAGE POLICY"}:
Please note that PUG REST is not designed for very large volumes (millions)
of requests. We ask that any script or application not make more than 5
requests per second, in order to avoid overloading the PubChem servers. If
you have a large data set that you need to compute with, please contact us
for help on optimizing your task, as there are likely more efficient ways to
approach such bulk queries.
}
\examples{
GetCid("aspirin", "name", quiet = TRUE)
}
