% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem.R
\name{GetCid}
\alias{GetCid}
\title{Match CID from other identifiers}
\usage{
GetCid(id, type = NULL, quiet = TRUE)
}
\arguments{
\item{id}{A vector of characters contains the identifiers of drugs you are
interested in.}

\item{type}{A character indicates the type of identifiers passed by \code{id}
argument. Available types are:
\itemize{
\item \strong{sid} The primary identifiers for the substance assigned by
PubChem.
\item \strong{smiles} Identifiers of drugs in the simplified molecular-input
line-entry system (SMILES)
\item \strong{inchi} The IUPAC International Chemical Identifier (InChI)
\item \strong{inchikey} Standard InChIKey of the drugs
\item \strong{name} Name for drugs. This type could be used for searching synonyms,
NCGC IDs, Chembl IDs, CAS or any other kind of identifiers.
}}

\item{quiet}{}
}
\description{
\code{GetCid} matches CIDs of drugs according to user-provided identifiers.
}
\details{
This function using the PubChem API
(\href{https://pubchem.ncbi.nlm.nih.gov/pug_rest/PUG_REST.html}{PUG REST}
to seach mached drugs. Available
identifiers are: name, SID, SMILES, InChI, SDF, InChIKey, molecula formula.

\strong{API USAGE POLICY:}
Please note that PUG REST is not designed for very large volumes (millions)
of requests. We ask that any script or application not make more than 5
requests per second, in order to avoid overloading the PubChem servers. If
you have a large data set that you need to compute with, please contact us
for help on optimizing your task, as there are likely more efficient ways to
approach such bulk queries.
}
