% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus.R
\name{GetDisease}
\alias{GetDisease}
\title{Extract the disease information of one cell line.}
\usage{
GetDisease(node)
}
\arguments{
\item{node}{An XMLInternalElementNode with only one cell line's information
which was extracted from Cellosaurus xml file.}
}
\value{
A data frame contains two variables: disease and disease_id
}
\description{
This function extract the cell line associated disease and disease ID (NCI
Thesaurus entry code) which the cell line associated with from an
XMLIntervalElementNode object extracted from Cellosaurus xml file. Combining
this function with \conde{\link[base]{apply}} or \code{\link[base]{sapply}}
can extract disease information from an XMLNodeSet object.
}
