% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_unichem.R
\name{GetIds}
\alias{GetIds}
\title{Get other identifiers from InChIKey}
\usage{
GetIds(inchikey)
}
\arguments{
\item{inchikey}{A vector of characters. It contains identifiers of the drug.}
}
\value{
A data frame with first column is the identifier passed to \code{ids}
argument and following columns are indentifiers of all resources integrated
by UniChem.
}
\description{
\code{GetIds} matches the identifiers of
\href{https://www.ebi.ac.uk/unichem/ucquery/listSources}{UniChem sources},
including DrugBank, KEGG, Zinc...
}
\details{
This function retrieves the of drug
via \href{https://www.ebi.ac.uk/unichem/info/webservices}{UniChem Web
Services}.
}
\examples{
aspirin <- GetIds(inchikey = "BSYNRYMUTXBXSQ-UHFFFAOYSA-N")
}
\author{
Jing Tang \email{jing.tang@helsinki.fi}
Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
}
