% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus.R
\name{GetNames}
\alias{GetNames}
\title{Extract primary name and synonyms of one cell line.}
\usage{
GetNames(node)
}
\arguments{
\item{node}{An \code{XMLInternalElementNode} with only one
\emph{cell-line-list} node extracted from \emph{Cellosaurus xml file}.}
}
\value{
A data frame contains two variables:
\itemize{
  \item \code{name} the primary name of cell line.
  \item \code{synonyms} synonyms of the cell line separated by semicolons.
}
}
\description{
\code{GetNames} extract primary name and synonyms from only one
\emph{cell-line-list} node in Cellosaurus xml file.
}
\details{
This function extracts the primary name (as "name") and synonyms (as
"synonyms") of an \code{\link[XML]{XMLIntervalElementNode}} object containing
one \emph{cell-line-list} node extracted from Cellosaurus xml file.

If you'd like to extract information from multiple \emph{cell-line-list}
nodes, combining this function with \code{\link[XML]{xpathSapply}} or
\code{\link[base]{sapply}} is recommanded.
}
\examples{
# parse the Cellosaurus xml file.
node <- GetAllCell(system.file("extdata",
                               "cellosaurus.xml",
                               package = "TidyComb"))

# extract "cell-line-list" nodes of "U251" and "U87" cell lines.
cell <- GetCell(node, "name", c("U251", "U87"))

# get first cell line (U251)'s primary name and synonyms
name <- GetNames(cell[[1]])

# get all cell lines' name and synonyms
names <- sapply(cell, GetNames)
}
