% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_pubchem.R
\name{GetPubNames}
\alias{GetPubNames}
\title{Get drug synonyms from PubChem}
\usage{
GetPubNames(cids, quiet = FALSE)
}
\arguments{
\item{cids}{A vector of characters or integers containing CID of drugs for
searching.}

\item{quiet}{A logical value. If it is \code{TRUE} the message for function
progress will be shutted down.}
}
\value{
A Data frame contains:
\itemize{
  \item \strong{cid} Inputted CIDs.
  \item \strong{name} The first name whose length is less than 40 characters
  in the synonyms list retrieved from PubChem.
  \item \strong{synonyms} Synonyms retrieved from PubChem
}
}
\description{
\code{GetPubNames} queries PubChem database with drug CIDs via
(\href{https://pubchem.ncbi.nlm.nih.gov/pug_rest/PUG_REST.html}{PUG REST} to
searching for synonyms of drug.
}
\examples{
GetPubNames("2244")
}
\author{
Jing Tang \email{jing.tang@helsinki.fi}
Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
}
