% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_cellosaurus.R
\name{GetTissue}
\alias{GetTissue}
\title{Extract the source tissue of one cell line.}
\usage{
GetTissue(node)
}
\arguments{
\item{node}{An XMLInternalElementNode with only one cell line's information
which was extracted from Cellosaurus xml file.}
}
\value{
A character. The tissue name of cell line according to CClE category.
}
\description{
This function extract source tissue according to cross-reference "CCLE Name"
from an XMLIntervalElementNode object extracted from Cellosaurus xml file.
Combining it with \code{apply} or \code{sapply} can extract tissue from an
XMLNodeSet object.
}
