% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cell_tissu_disease.R
\name{MatchCellAcc}
\alias{MatchCellAcc}
\title{Match cell line name with cellosaurus accession}
\usage{
MatchCellAcc(name)
}
\arguments{
\item{name}{A vector of charactors contains names of interested cell lines}
}
\value{
A data frame contains:
\itemize{
  \item \strong{input_name} The input cell lines names
  \item \strong{cellosaurus_accession} The cellosaurus accession matched with
  cell line names.
  \item \strong{all_name} All names for the mached cell lines, including
  primary name and synonyms.
}
}
\description{
Match cell line name with cellosaurus accession
}
