% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyComb.R
\docType{package}
\name{TidyComb}
\alias{TidyComb}
\title{TidyComb: Preparing drug combination data for DrugComb}
\description{
This package wrapped the functions that could be used to tidy drug
combination data from various sources into the format that is required by
uploading to DrugComb database.
}
\details{
The TidyComb package provides three categories of important
functions :
\enumerate{
  \item Retrieving data from external databases, like PubChem, Chembl, ...;
  \item Estimating cell line's sensitivity to drugs. Methods includs fitting
  and visaulizing dose-response curve, ;
  \item Calculating drug combination effect surface.
}
}
\section{Retrieving data from databases}{

The following functions are provided to
}

